Remove Home composable and rename file
